Remove unnecessary produce from immer, and check if there is a cycle in the queries structure
